started renaming all the index.js with Products4Sale
